+++ modules/account/index.ts
@@ -1,3 +1,4 @@
 export { getAccountLinksEnabled, useAccountLinksEnabled } from "./core/accountLinksConfig";
-export { getAccountCabinetUrl, openAccountCabinet } from "./core/openAccountCabinet";
+export { getAccountCabinetUrl, openAccountCabinet, type CabinetContext } from "./core/openAccountCabinet";
+export { resolveBillingCurrency, type BillingCurrency } from "./core/billingCurrency";
 export { MembershipEventsBridge } from "./ui/MembershipEventsBridge";